Terminology updates to README.md
Applied terminology adjustments, with regards to "install" and "deploy", and "SAS Viya" and "SAS Viya platform". Removed the "SAS Viya" illustration at the beginning of the README. Also added a new heading called "Using the Monitoring Components" and moved the "Access the Dashboards" and "Adding Grafana Alert Rules" sections there as these sections are about using the software rather than deploying it.